Done. I implemented some logic that won´t allow the user to create a contact if it there are no categories existent. It solves the problem of the program "blowing up" in that moment.
I also considered creating a default "category" object that would basically work as a "null" if the user didnt create any objects yet.
I'm not sure, but i think that way it would enable the user to create contacts with "no category" (default) if no categories were yet created by the user.

However, i decided to simply not allow the user to create any contacts at all. As I though, thats the way the app is supposed to work.
-------------
Also, I made some changes that will print a message if the user tries to update or delete (contacts or categories) and there are none existant. Instead of just "blowing up".

+++ PhoneBookCarDioLogics/PhoneBookCarDioLogics/Controllers/ContactsController.cs
@@ -9,26 +9,36 @@ internal class ContactsController
 {
     internal static void InsertContact()
     {
-        var contact = new Contact();
-        contact.Name = AnsiConsole.Ask<string>("Contact's Name:");
-        contact.PhoneNumber = AnsiConsole.Ask<string>("Contact's phone Number:");
-        contact.CategoryID = CategoryController.GetCategoryOptionInput().CategoryId;
-        contact.Email = AnsiConsole.Confirm("Add email to contact?") ? AnsiConsole.Ask<string>("Contact's email:") : contact.Email = "";
-
-        if (contact.Email != "")
-        {
-            string emailDomain = contact.Email.Split('@')[1];
-            if (emailDomain != "gmail.com" && emailDomain != "hotmail.com")
-            {
-                contact.Email = "";
-                Console.WriteLine("Email domain not supported");
-                Console.ReadLine();
-            }
-        }
-
         using var context = new PhonebookAppDbContext();
-        context.Add(contact);
-        context.SaveChanges();
+
+        if (context.Categories.ToList().Count() > 0) //if there are categories the code below here runs
+        {
+
+            var contact = new Contact();
+            contact.Name = AnsiConsole.Ask<string>("Contact's Name:");
+            contact.PhoneNumber = AnsiConsole.Ask<string>("Contact's phone Number:");
+            contact.CategoryID = CategoryController.GetCategoryOptionInput().CategoryId;
+            contact.Email = AnsiConsole.Confirm("Add email to contact?") ? AnsiConsole.Ask<string>("Contact's email:") : contact.Email = "";
+
+            if (contact.Email != "")
+            {
+                string emailDomain = contact.Email.Split('@')[1];
+                if (emailDomain != "gmail.com" && emailDomain != "hotmail.com")
+                {
+                    contact.Email = "";
+                    Console.WriteLine("Email domain not supported");
+                    Console.ReadLine();
+                }
+            }
+
+            context.Add(contact);
+            context.SaveChanges();
+        }
+        else
+        {
+            Console.WriteLine("Can't create a contact! No categories exist yet!");
+            Console.ReadLine();
+        }
     }
 
     internal static List<Contact> ShowContacts()
@@ -46,36 +56,65 @@ internal class ContactsController
     internal static void ChangeContact()
     {
         var contact = GetContactOptionInput();
-        contact.Name = AnsiConsole.Confirm("Update name?") ? AnsiConsole.Ask<string>("Contact's new name:") : contact.Name;
-        contact.PhoneNumber = AnsiConsole.Confirm("Update phone number?") ? AnsiConsole.Ask<string>("Contact's new phone number:") : contact.PhoneNumber;
-        contact.Email = AnsiConsole.Confirm("Update email?") ? AnsiConsole.Ask<string>("Contact's new email:") : contact.Email;
-        contact.Category = AnsiConsole.Confirm("Update category?") ? CategoryController.GetCategoryOptionInput() : contact.Category;
 
-        using var db = new PhonebookAppDbContext();
-        db.Update(contact);
-        db.SaveChanges();
+        if (contact != null)
+        {
+            contact.Name = AnsiConsole.Confirm("Update name?") ? AnsiConsole.Ask<string>("Contact's new name:") : contact.Name;
+            contact.PhoneNumber = AnsiConsole.Confirm("Update phone number?") ? AnsiConsole.Ask<string>("Contact's new phone number:") : contact.PhoneNumber;
+            contact.Email = AnsiConsole.Confirm("Update email?") ? AnsiConsole.Ask<string>("Contact's new email:") : contact.Email;
+            contact.Category = AnsiConsole.Confirm("Update category?") ? CategoryController.GetCategoryOptionInput() : contact.Category;
+
+            using var db = new PhonebookAppDbContext();
+            db.Update(contact);
+            db.SaveChanges();
+        }
+        else 
+        {
+            Console.WriteLine("No contacts exist yet!");
+            Console.ReadLine();
+        }
     }
 
     internal static void RemoveContact()
     {
         var contact = GetContactOptionInput();
-
-        using var db = new PhonebookAppDbContext();
-        db.Remove(contact);
-        db.SaveChanges();
+        
+        if (contact != null)
+        {
+            using var db = new PhonebookAppDbContext();
+            db.Remove(contact);
+            db.SaveChanges();
+        }
+        else
+        {
+            Console.WriteLine("No contacts exist yet!");
+            Console.ReadLine();
+        }
     }
 
     private static Contact GetContactOptionInput()
     {
-        var contacts = ShowContacts();
-        var contactsArray = contacts.Select(x => x.Name).ToArray();
-        var option = AnsiConsole.Prompt(new SelectionPrompt<string>()
-            .Title("Choose contact")
-            .AddChoices(contactsArray));
-        var id = contacts.Single(x => x.Name == option).ContactId;
-        var contact = GetContactByID(id);
+        List<Contact> contacts = ShowContacts();
+        bool contactExists;
 
-        return contact;
+        if(contacts.Count > 0)
+        {
+            var contactsArray = contacts.Select(x => x.Name).ToArray();
+
+            var option = AnsiConsole.Prompt(new SelectionPrompt<string>()
+                .Title("Choose contact")
+                .AddChoices(contactsArray));
+            var id = contacts.Single(x => x.Name == option).ContactId;
+            var contact = GetContactByID(id);
+
+            contactExists = true;
+
+            return contact;
+        }
+        else
+        {
+            return null;
+        }
     }
 
     private static Contact GetContactByID(int id)
@@ -94,34 +133,42 @@ internal class ContactsController
         Console.WriteLine("Select who do you want to send an email to:");
         var contact = GetContactOptionInput();
 
-        string emailRecipient = contact.Email;
-
-        if (emailRecipient != "")
+        if(contact != null)
         {
-            string emailSender = AnsiConsole.Ask<string>("Sender's email:");
-            string emailSenderPassword = AnsiConsole.Prompt(new TextPrompt<string>("Sender's email password:").Secret());
+            string emailRecipient = contact.Email;
 
-
-            string emailDomain = emailSender.Split('@')[1];
-
-            SmtpClient smtpClient = EmailService.SetEmailClient(emailDomain, emailSender, emailSenderPassword);
-
-            //if the email domain is supported. The user is allowed to create and send the email.
-            if (emailDomain == "gmail.com" || emailDomain == "hotmail.com")
+            if (emailRecipient != "")
             {
-                MailMessage mailMessage = new MailMessage(emailSender, emailRecipient)
-                {
-                    Subject = AnsiConsole.Ask<string>("Subject of the email:"),
-                    Body = AnsiConsole.Ask<string>("Write the content of the email:"),
-                    IsBodyHtml = false, // Set to true if you want to use HTML in the email body
-                };
+                string emailSender = AnsiConsole.Ask<string>("Sender's email:");
+                string emailSenderPassword = AnsiConsole.Prompt(new TextPrompt<string>("Sender's email password:").Secret());
 
-                smtpClient.Send(mailMessage);
+
+                string emailDomain = emailSender.Split('@')[1];
+
+                SmtpClient smtpClient = EmailService.SetEmailClient(emailDomain, emailSender, emailSenderPassword);
+
+                //if the email domain is supported. The user is allowed to create and send the email.
+                if (emailDomain == "gmail.com" || emailDomain == "hotmail.com")
+                {
+                    MailMessage mailMessage = new MailMessage(emailSender, emailRecipient)
+                    {
+                        Subject = AnsiConsole.Ask<string>("Subject of the email:"),
+                        Body = AnsiConsole.Ask<string>("Write the content of the email:"),
+                        IsBodyHtml = false, // Set to true if you want to use HTML in the email body
+                    };
+
+                    smtpClient.Send(mailMessage);
+                }
+            }
+            else
+            {
+                Console.WriteLine("This contact does not have a valid email adress.");
+                Console.ReadLine();
             }
         }
         else
         {
-            Console.WriteLine("This contact does not have a valid email adress.");
+            Console.WriteLine("Can´t send email! No contacts exist yet!");
             Console.ReadLine();
         }
     }
